+++ test/error/binding_not_found_param.sv
@@ -1,4 +1,4 @@
-// pattern: unknown binding "R" specified for parameter overrides in instance "e" of "example"
+// pattern: unknown binding "R" specified for parameter overrides in instance "e" of module "example", 2 available \("P", "Q"\)
 module example;
     parameter P = 1;
     parameter Q = 1;