+++ test/error/binding_overflow_class.sv
@@ -1,4 +1,4 @@
-// pattern: too many bindings specified for parameters in class specialization of "example"
+// pattern: too many bindings specified for parameters in class specialization of "example": 3 specified \(1, 2, 3\), but only 2 available \("P", "Q"\)
 class example #(
     parameter P = 1,
     parameter Q = 1