+++ test/error/binding_overflow_port.sv
@@ -1,4 +1,4 @@
-// pattern: too many bindings specified for port connections in instance "e" of "example"
+// pattern: too many bindings specified for port connections in instance "e" of module "example": 3 specified \(1'b1, 1'b0, 1'b0\), but only 2 available \("x", "y"\)
 module example(
     input x, y
 );