+++ test/error/interface_param_mismatch_expr.sv
@@ -1,4 +1,4 @@
-// pattern: parameter "P" in instance "intf" of "Interface" expects an expression, but was given type logic
+// pattern: parameter "P" in instance "intf" of interface "Interface" expects an expression, but was given type logic
 interface Interface;
     parameter P = 0;
     logic [P-1:0] x;