+++ test/error/interface_param_mismatch_type.sv
@@ -1,4 +1,4 @@
-// pattern: parameter "P" in instance "intf" of "Interface" expects a type, but was given expression 1
+// pattern: parameter "P" in instance "intf" of interface "Interface" expects a type, but was given expression 1
 interface Interface;
     parameter type P;
     P x;